double precision implemented

+++ DiGi.ComputeSharp/Core/Constans/Tolerance.cs
@@ -5,16 +5,16 @@
         /// <summary>
         /// Dictance tolerance.
         /// </summary>
-        public const float Distance = 1e-6f;
+        public const double Distance = 1e-6;
 
         /// <summary>
         /// Macro dictance tolerance.
         /// </summary>
-        public const float MacroDistance = 1e-3f;
+        public const double MacroDistance = 1e-3;
 
         /// <summary>
         /// Angle tolerance. Equivalent of 2 degrees
         /// </summary>
-        public const float Angle = 0.0349066f;
+        public const double Angle = 0.0349066;
     }
 }
