+++ DiGi.ComputeSharp/Spatial/Create/Line3Intersections.cs
@@ -6,7 +6,7 @@ namespace DiGi.ComputeSharp.Spatial
     public static partial class Create
     {
 
-        public static IEnumerable<Line3Intersection> Line3Intersections(Line3 line, IEnumerable<Line3> lines, float tolerance)
+        public static IEnumerable<Line3Intersection> Line3Intersections(Line3 line, IEnumerable<Line3> lines, double tolerance)
         {
             if (line.IsNaN() || lines == null || lines.Count() == 0)
             {
